Fix bug where player would get stuck in wall
While moving between rooms the player would get stuck in a wall in the opposite
side of the room. After some inspection I saw that it had to do with the new
room not spawning and just being translated on top of the player.

My guess is it has to do with the player-room-min function, which controlled
where the room got translated and how collision was detected, was translating
the room and updating where collision with the wall were taking place before a
new room could be spawned.

This is fixed by using the same value, bounds-x and bounds-y, for the room
translation and collision and room spawning.